requirements.txt
asgiref==3.4.1
click==8.0.3
Django==4.0
djangorestframework==3.13.1
h11==0.12.0
pytz==2021.3
sqlparse==0.4.2
uvicorn==0.16.0
whitenoise==5.3.0
PyJWT==2.3.0
django-cors-headers==3.10.1